src/main/java/StringHelper.java
/**
 * TODO https://github.com/in28minutes/JUnitIn28Minutes
 * https://www.youtube.com/watch?v=o5k9NOR9lrI&ab_channel=in28minutes-GetCloudCertified
 */
public class StringHelper {
//Remove 1st 2 A from String AACD-->CD
	public String truncateAInFirst2Positions(String str) {
		if (str.length() <= 2)
			return str.replaceAll("A", "");

		String first2Chars = str.substring(0, 2);
		String stringMinusFirst2Chars = str.substring(2);

		return first2Chars.replaceAll("A", "") 
				+ stringMinusFirst2Chars;
	}

	public boolean areFirstAndLastTwoCharactersTheSame(String str) {

		if (str.length() <= 1)
			return false;
		if (str.length() == 2)
			return true;

		String first2Chars = str.substring(0, 2);

		String last2Chars = str.substring(str.length() - 2);

		return first2Chars.equals(last2Chars);
	}

}